+++ packages/go_router_builder/lib/src/route_config.dart
@@ -276,7 +276,9 @@ class GoRouteConfig extends RouteBaseConfig {
         );
       }
     }
-    final String fromStateExpression = decodeParameter(element, _pathParams);
+    final List<ElementAnnotation>? metadata = _fieldMedatada(element.name);
+    final String fromStateExpression =
+        decodeParameter(element, _pathParams, metadata);
 
     if (element.isPositional) {
       return '$fromStateExpression,';
@@ -301,7 +303,8 @@ class GoRouteConfig extends RouteBaseConfig {
       );
     }
 
-    return encodeField(field);
+    final List<ElementAnnotation>? metadata = _fieldMedatada(fieldName);
+    return encodeField(field, metadata);
   }
 
   String get _locationQueryParams {
@@ -700,6 +703,10 @@ $routeDataClassName.$dataConvertionFunctionName(
 
   PropertyAccessorElement? _field(String name) =>
       routeDataClass.getGetter(name);
+
+  List<ElementAnnotation>? _fieldMedatada(String name) => routeDataClass.fields
+      .firstWhereOrNull((FieldElement element) => element.displayName == name)
+      ?.metadata;
 
   /// The name of `RouteData` subclass this configuration represents.
   @protected